refactor: update login and logout flows with 2FA support and token generation improvements
- Adjusted `LoginUserService` to handle 2FA verification flow via the login handler and streamlined token generation logic.
- Updated `LoginSessionCommandHandler` to generate tokens inline and return 2FA-required responses when applicable.
- Refactored `LogoutSessionCommandHandler` to implement `BaseHandlerCommand` with a `void` result type.
- Modified `LoginUserViewModel` and `GetUserViewModel` to properly handle `twoFactorEnabled` and `token` fields with nullability refinements.
