+++ user-service/src/Application/ViewModels/LoginUserViewModel.ts
@@ -1,6 +1,6 @@
 export class LoginUserViewModel
 {
-    constructor(public token: string, public uuid: string | null = null,
+    constructor(public token: string | null, public uuid: string | null = null,
                 public username: string | null = null, public profilePic: string | null = null){}
 
 }
