+++ user-service/src/Domain/Command/Handlers/LoginSessionCommandHandler.ts
@@ -3,21 +3,51 @@ import {UserRepository} from "../../../Infrastructure/Persistence/Repositories/C
 import {NotificationError} from "../../../Shared/Errors/NotificationError.js";
 import {UserSessionCommand} from "../CommandObject/UserSessionCommand.js";
 import {PasswordHashVO} from "../../ValueObjects/PasswordHashVO.js";
+import {LoginUserViewModel} from "../../../Application/ViewModels/LoginUserViewModel.js";
+import {FastifyReply, FastifyRequest} from "fastify";
+import {UserSessionDTO} from "../../../Application/DTO/ToCommand/UserSessionDTO.js";
+import {User} from "../../Entities/Concrete/User.js";
 
-export class LoginSessionCommandHandler implements BaseHandlerCommand<UserSessionCommand>
+export class LoginSessionCommandHandler implements BaseHandlerCommand<UserSessionCommand, LoginUserViewModel>
 {
    private UserRepository: UserRepository;
 
-   constructor(userRepository: UserRepository, notification: NotificationError) {
-       this.UserRepository = userRepository;
+   constructor(userRepository: UserRepository, notification: NotificationError)
+   {
+      this.UserRepository = userRepository;
    }
 
-   async Handle(command: UserSessionCommand) : Promise<void>
+   async Handle(command: UserSessionCommand, request: FastifyRequest<{ Body: UserSessionDTO}>, reply: FastifyReply): Promise<LoginUserViewModel>
    {
       const user = await this.UserRepository.GetUserEntityByEmail(command.Email);
 
       user?.ChangeStatusOnline(command.isOnline);
 
       await this.UserRepository.Update(user!.Uuid, user);
+
+      return this.GenerateToken(user, reply, request);
+   }
+
+   private GenerateToken(user: User | null, reply: FastifyReply, request: FastifyRequest<{ Body: UserSessionDTO }>)
+   {
+      const body = request.body;
+
+      if (user && user.TwoFactorEnabled)
+          return new LoginUserViewModel(null, user.Uuid, user.Username, user.ProfilePic);
+
+      const token = request.server.jwt.sign({
+         email: body.email,
+         isAuthenticated: true,
+      }, {expiresIn: '1d'});
+
+      reply.setCookie('token', token, {
+         httpOnly: true,
+         secure: false, // Deixei como false porque a aplicação ainda está em HTTP, mas deve ser true em HTTPS
+         sameSite: 'lax',
+         path: '/',
+      });
+
+      //TODO: update esse cara debaixo
+      return new LoginUserViewModel(token);
    }
 }
